Refactored hook into its own file, used css variables for color
